po-*: move the Python goat exercise to the gimp30-std-plug-ins domain.
Since commit ecbc38f9e9, the Python goat exercise was tracked by the
po-python/ Gettext files. This was an error as all its strings are
common to all other goat exercise plug-ins, so it makes no sense to
separate them (it only duplicates the work).

Actually separating the translation domains by plug-in language was a
bad idea from the start. Why do we care that a translatable string is
written in this or that language? All our plug-ins should just share a
common text domain. This way, we avoid useless duplication of common
strings.