NEWS: update with new "file-heif-av1-load" procedure.
`gimp-2-10` branch has a common load procedure for HEIC and AVIF, but
the main file-heif contributor (Daniel Novomesky) considers it better to
have 2 procedures. Among other things, it helps to allow using a
separate plug-in for AVIF while this procedure is still used for HEIC.

This won't be backported to 2.10 because it changes the existing
procedure usage. So it's a master-only change, hence adding it to NEWS.