build: run `crossroad update` before installing each dependency set.
Lately we have a bunch of CI failure because of package not found. I am
unsure of the exact problem yet, hence this first attempt of a possible
solution. The idea is that *maybe* the repository listing is already
outdated (i.e. between dependencies installation for babl, then GEGL,
then GIMP). I am really not sure this is the core issue, but this is a
test for now.