plug-ins: improve error handling and use GError in metadata-editor/viewer.
metadata-editor and metadata-viewer did return GIMP_PDB_SUCCESS when
loading the dialog ui failed, we changed this to GIMP_PDB_EXECUTION_ERROR
and now use a GError to report it back to GIMP.

In metadata-editor also improve the error handling when loading the calendar
dialog fails.

The error also only showed in the console so change that to show a dialog
or the error console.

(cherry picked from commit 0ffee001bf1dfbd97ad866bc5c157c59d3433a9e)

# Conflicts:
#	plug-ins/metadata/metadata-editor.c
#	plug-ins/metadata/metadata-viewer.c